+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ### Fixed
 - CP notifications were not being displayed after field actions occurred
 - Instance where an environment 
+- Sync actions and transformations
 
 ## 1.0.0-beta.1 - 2018-04-26
 ### Fixed